[REFACTOR] Resolve Merge Issue
  - Resolve structure-related issue occurred after merge.
