+++ feature/main/src/main/java/com/doyoonkim/main/contract/NoticeSearchContract.kt
@@ -3,15 +3,21 @@ package com.doyoonkim.main.contract
 import com.doyoonkim.common.base.UiEvent
 import com.doyoonkim.common.base.UiSideEffect
 import com.doyoonkim.common.base.UiState
+import com.doyoonkim.model.BookmarkAsListElementVO
 import com.doyoonkim.model.NoticeVO
 
 data class NoticeSearchState(
     val searchKeyword: String = "",
+    val fetchingSource: FetchingSource = FetchingSource.REMOTE,
     val isError: Boolean = false,
     val isFetching: Boolean = false,
+    val isSearchResultEmpty: Boolean = false,
     val canRequestMoreNotices: Boolean = true,
-    val fetchResult: List<NoticeVO> = emptyList()
+    val fetchResult: List<NoticeVO> = emptyList(),
+    val localFetchResult: List<BookmarkAsListElementVO> = emptyList()
 ) : UiState
+
+enum class FetchingSource { REMOTE, LOCAL }
 
 sealed class NoticeSearchEvent : UiEvent {
     data class InitiateSearch(val keyword: String): NoticeSearchEvent()
